Add a Nix devshell
Let's try Haskell, it's supposed to be good for writing compilers after
all. (It is quite nice, although in my experience it's not for writing
parsers if you want a human-friendly parser.) And Haskell offers a lot
of tools to be confident that the compiler itself is correct. It should
be possible to translate into Idris later if needed.