Emit some output from the "compiler"
It can generate constraints and definitions for the circuit language
now. I needed to make Circuit be Monad, not just Applicative, which on
the one hand is bad because this opens up the control flow can of worms.
But on the other hand, it does enable many nice things as well. For
example, I can put a fresh variable counter in there, instead of having
to infer it.