Add constraint canonicalization
It generates slightly fewer constraints now. And I have some ideas on
how we could optimize the contraints further: if we have one of the form

  (a1 r1 + a2 r2 + ... + an rn) * 1 = r(n+1)

Then we can replace occurrences of r(n+1) with its definition. It's a
bit tedious but it shouldn't be too difficult, we make a map of all such
definitions, then in a second pass we apply them.