Apply optimizer to constraints
Maybe I should have invested a bit more in not generating the silly
constraints in the first place. But this works! And it's more general.
Let's go with it for now. In the simpler form, I can actually recognize
where the constraints come from, but the signs don't always make sense
... did I switch a Sub somewhere?